@@ -2,6 +2,10 @@
 #
 # SPDX-License-Identifier: MIT
 
-__version__ = "0.3.0"
+"""Main __init__ file."""
+
+__version__ = "0.5.0"
 
 from .sphinx_datatables import setup
+
+__all__ = ["setup"]
